bluetooth: Add support of GATT characteristic read value
This patches includes the following changes.
- In GattInterfaceHandler, a map is added to track the D-Bus requests
and the response to reply to. Once HandleCharacteristicReadValue() is
triggered, we add an entry in the map and call Gatt with
OnReadCharacteristicValue() which will later be invoked.
- In Gatt, we add ReadCharacteristicValue() as the core API for reading
characteristic value. It calls Newblue with OnGattClientReadValue()
which will later be invoked. Note that OnGattClientReadValue() is shared
between characteristic and descriptor, so we add extra request type to
tell them apart.
- In Newblue, we add GattClientReadValue() which calls gattClientRead()
to perform read. The caller of GattClientReadValue() will get notified
on GattClientReadValueCallback().

Apart from the above call chain, this also adds some enum classes to
represent error code and types for operation/request tracking.

BUG=b:134425062
TEST=run bluetoothctl --service=org.chromium.Newblue to perform
(1) Scan, pair and connect to a BLE mouse. Gatt browsing started right
after connections, so verify the attributes are printed and
ServicesResolved property became true.
(2) Run "list-attributes" and see the expected attributes printed.
(3) Run "select-attributes" to set the target attribute to read. We
choose battery level characteristic in this case.
(4) Run "read" and verify that the value matches with the result with
BlueZ.

Change-Id: Ife77be8f7231cc2e24ebcad114f3506f7eb8719f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1756936
Tested-by: Miao-chen Chou <mcchou@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Yoni Shavit <yshavit@chromium.org>