cups_proxy: start when starting concierge
Cups proxy does not need concierge services, so we do not need to wait
for it to start. In fact, we want it to be already started when
concierge runs, because we expect all services that concierge/VMs
depend on to be started so that concierge can service requests as soon
as it is running.

BUG=b:117218264
TEST=Start Plugin VM

Change-Id: I5d0d2f6aea1bee24aff97a5485290a801d72724a
Reviewed-on: https://chromium-review.googlesource.com/1761668
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Ready: Dmitry Torokhov <dtor@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>