cups_proxy: create runtime directory with right permissions
Make sure we create runtime directory and give it correct ownership when
starting the proxy. This is needed to make sure we'll end up with
correct ownership in cases when concierge starts first and it creates
the proxy runtime directory (concierge has to try to create it because
it bind-mounts it into the jails).

BUG=b:117218264
TEST=Start Plugin VM

Change-Id: I00072b21f9bf68f53a5e45f9d860395cde59f574
Reviewed-on: https://chromium-review.googlesource.com/1762056
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Ready: Dmitry Torokhov <dtor@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>