bluetooth: Add support of GATT descriptor read value
This patches includes the following changes.
- In GattInterfaceHandler, a map is added to track the D-Bus requests
and the response to reply to. Once HandleDescriptorReadValue() is
triggered, we add an entry in the map and call Gatt with
OnReadDescriptorValue() which will later be invoked.
- In Gatt, we add ReadDescriptorValue() as the core API for reading
descriptor value. It calls Newblue with OnGattClientReadValue()
which will later be invoked.

BUG=b:134425062
TEST=run bluetoothctl --service=org.chromium.Newblue to perform
(1) Scan, pair and connect to a BLE keyboard. Gatt browsing started right
after connections, so verify the attributes are printed and
ServicesResolved property became true.
(2) Run "list-attributes" and see the expected attributes printed.
(3) Run "select-attributes" to set the target attribute to read. We
choose Client Characteristic Configuration descriptor in this case.
(4) Run "read" and verify that the value matches with the result with
BlueZ.

Change-Id: I734f52ebec03dddcc1116d64c29ae06c97944431
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1761840
Tested-by: Miao-chen Chou <mcchou@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Yu Liu <yudiliu@chromium.org>