shill: service: Fix UserInitiated{Connect,Disconnect} semantics
Prior to this change, there was an issue in which a UserInitiatedConnect that
fails quickly would result in unnecessarily setting |is_in_user_connect_|. For
example, if a client tried to connect to a Service that was not connectable(),
the failure would not be recorded by ReportUserInitiatedConnectionResult and the
next connection (which would likely be an AutoConnect triggered by
SetConnectableFull) would be considered a user initiated connection.

A similar, but latent, issue is fixed for UserInitiatedDisconnect. Since
Cellular is the only technology that uses SetFailureSilent to drive a Service
state back to kStateIdle during a disconnect, but also happens to have an
asynchronous disconnect in which SetFailureSilent is called *after*
UserInitiatedDisconnect has completed, this issue is currently not hit by any
flows.

BUG=None
TEST=-All unit tests are passing.
     -Manual testing with connections and disconnections through the
      Chrome UI.
     -Manual testing involving LOGs in UserInitiatedConnect and
      `connect-service` calls (as the Chrome UI doesn't let us do
      things like connect to a connected Service).

Change-Id: Icbe14cc96e20b3ae92d32fd64c2732d625deb7a9
Reviewed-on: https://chromium-review.googlesource.com/1762627
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>