shill: Add mockable keyword
In general, the virtual keyword is used for methods that are going to
be mocked, even if those methods should not be inherited by non-mock
children. This works fine for projects that have a flat or very simple
inheritance hierarchy. Shill, however, has neither of those, and the
most complex inheritance hierarchies (e.g. the Service and Device
hierarchies) become much more difficult to read and understand when
one is never sure if a virtual method is used just for mocks or is
also overridden by children.

This change creates a separate 'mockable' keyword for methods that
should only be overridden by mocks. The 'mockable' behavior is
enforced by being #defined to virtual only in test builds.

BUG=chromium:997295
TEST=None.

Change-Id: I0800811fbeadf6c61895cc410deec05770982c94
Reviewed-on: https://chromium-review.googlesource.com/1768862
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>