shill: vpn: Stop overridding SetConnection
CL:1748120 introduced a regression by triggering a latent infinite
loop that was not previously reachable because the previous
SetConnection override did not work as "intended".

Ever since it was introduced in CL:22461, the
VPNService::SetConnection override essentially behaved as if it wasn't
there at all. Since Device::DestroyConnection calls
Service::SetConnection(nullptr) prior to destroying the Connection
instance, and since rebinding a Binder does *not* cause--nor has it
ever caused--the client_disconnect_callback to be called, the Binder
in VPNService::SetConnection was never bound to a Connection during
the destruction of the Connection.

This change simply removes the override of SetConnection, as the
previous implementation essentially behaved as if that was the case.

BUG=chromium:996957
TEST=-All unit tests are passing.
     -Triggering a VPN disconnect after successfully connecting does
      not cause VPNDriver::OnConnectionDisconnected to be called (as
      was the behavior before CL:1748120). With l2tp, this is verified with:
        `cat /var/log/net.log | grep Underlying`

Change-Id: Ief557c9b429dd55bedffc3aefaec4f1b6e6db680
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1769576
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>