camera: camera3_test: Fix Aborted (core dumped) issue
The main reason for this bug is because there is no release hal
so before the end of the google ut process. It is to release hal
so after the end of the google ut process. The static storage
area that causes hal so is released in advance. At this time,
hal and then use the static zone variable will have a crash problem.
Modify the google ut code. Let it release hal so before the main function ends.

BUG=b:140156591
TEST=cros_camera_test  --camera_hal_path=/usr/lib64/camera_hal/intel-ipu6.so
--gtest_filter=Camera3ModuleFixture.NumberOfCameras & All google ut cases tests
on the ipu3 platform passed.

Change-Id: Iddd4fb9f6faee85b9faab7c18f035ac16ce2471a
Signed-off-by: shigangx.yang <shigangx.yang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1769101
Tested-by: Tianshu Qiu <tian.shu.qiu@intel.com>
Commit-Queue: Tianshu Qiu <tian.shu.qiu@intel.com>
Commit-Queue: Shik Chen <shik@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>