shill: Make DefaultServiceObserver and pass logical and physical Service
When Manager finishes sorting its Services, it notifies clients (any
instance that has called AddDefaultServiceObserver) of the
primary physical Service, if it has changed. In order to implement the
concept of a "lower Device" for VirtualDevices, clients must also be
able to receive notifications when the logical Service (i.e. the
"default Service") has changed. For example, if a VPNService goes down
but none of the physical Services change, the physical_service will be
unchanged, but the logical_service will change from whatever
VPNService was active to the physical_service. Virtual interfaces
(e.g. a Crostini tun interface using the default Service) will care
about the change of either physical or logical Service.

The OpenVPN and Chrome VPNDrivers are modified to behave as before,
despite the potential for a notification where the physical_service
has not changed.

BUG=chromium:999589
TEST=-All unit tests are passing.
     -Manually tested connected to a VPN, changing back and forth
      between underlying physical networks, and then disconnecting the
      VPN.

Change-Id: Ia554cb8f771406b981c06884aafa9459ca957098
Reviewed-on: https://chromium-review.googlesource.com/1803754
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>