appfuse: Replace WatchFileDescriptor by FileDescriptorWatcher.
This is preparation of the libchrome uprev.
MessageLoopForIO::WatchFileDescriptor is removed in the next uprev,
and the replacement is base::FileDescriptorWatcher.

This CL replaces it.
Because the previous implementation unconditionally watches
"writable" state, it was busy loop.
This CL fixes it by watching writable state only if it is necessary.

BUG=chromium:909719
TEST=Ran cheets_AppfuseTest_P.

Change-Id: I19be9c51ec82937243caee93d84968e23a5f7ef9
Reviewed-on: https://chromium-review.googlesource.com/1772941
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Ready: Hidehiko Abe <hidehiko@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>