power: use event_count instead of wakeup_count
event_count increments for a given device when pm_wakeup_event() is
called by the associated driver. wakeup_count on the other hand
increments only for the first device that reported the wake when in
suspend/resume path.

This is a problem in couple of scenarios:
  1. When a wake is triggered by USB hid, wakeup count might only
     increments for root hub controller and not for the actual hid.
  2. When a wake is triggered by a mfd device, only the mfd device
     wakeup count might increment and the sub-platform device's (enumerated
     by mfd devices) wakeup count might not increment.

This change thus changes powerd code to depend on
event_count rather than wakeup_count.

Please look at
https://lore.kernel.org/linux-pm/20190619175142.237794-1-ravisadineni@chromium.org/
for more info on why wakeup_count is not a reliable.

BUG=b:133262012, b:65976859
TEST=Deploy on nocturne on test dark and full wake sources.

Cq-Depend: chromium:1771193, chromium:1767049
Change-Id: I4c84ed14d1fe4faf3f8643bc521487b4c83a728b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1775601
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>