cryptohome: Move unmount operations to the MountHelper class.
With all mount operations moved to the helper, this CL now moves
all unmount operations to the helper as well. Having both mount and
unmount operations in the helper is a prerequisite to moving these
operations out of process and eventually to carry them out in a
non-root mount namespace.

While the actual unmount operations and other clean-up functionality
got moved to the MountHelper class, this functionality is still
triggered in and by the Mount class. When moving the functionality
out of process, more of the unmount and clean-up logic will be
triggered by the MountHelper class.

This CL also removes the RememberMount/RememberBind functions, which
were kept around for testing in a previous CL.

BUG=chromium:985492
TEST=Unit tests pass.
TEST=Add new user to Chromebook successfully.
TEST=Log in, create local data, log out. No cryptohome errors.
TEST=Log back in, local data is still present.
TEST=Log into Guest mode, log out. No cryptohome errors.

Change-Id: I9c53fda4134bd93b7c31ec7625c847db2ae4151f
Reviewed-on: https://chromium-review.googlesource.com/1783157
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Ready: Jorge Lucangeli Obes <jorgelo@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>