dlcservice: Convert check update_engine status to a getter
Dlcservice should only be allowed to install when update_engine is IDLE
(which means there is no install or update in progress and no update
has been applied).

Dlcservice should only be allowed to uninstall when update_engine is
IDLE or UPDATED_NEED_REBOOT. It is still valid to uninstall when there
has been an update applied, because it will clear the /var/cache to
which the update is applied to.

This conversion is to provide a finer grain of control in error
propagation for each update_engine status. It is clear to see how
dlcservice should act for each update_engine status.

BUG=chromium:987287
TEST=FEATURES="test" emerge-$B dlcservice
TEST=tast run $TEST_IP platform.DLCService

Change-Id: Ibd3932fd99c9b4ba32909cfce0e4c5298ed922ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1745826
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>