system_api: Add patchpanel dbus service.
This patch introduces a DBus interface to arc-networkd that will be
initially used by Concierge for obtaining TAP devices for ARCVM and
Crostini.

It is specifically not called "ArcNetworkd" since the stack is less and
less ARC-specific and more and more general purpose guest networking
support.

BUG=b:133707144
TEST='emerge-$BOARD system_api' also as part of depedent patch chain,
verified that the service works as expected.

Change-Id: Ie0614a89ab8a0f2bdf36fbfc0643b25b929dfd78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1885531
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>