attestation/tpm_manager: Fix ODR violation in protobufs
attestation/tpm_manager protos in system_api is used in both lib*.so and the
actual binary/test runners. The end result is that when the actual binary/
test runners load the lib*.so, there'll be 2 copies of the protos. This is
solved by not linking the protos into lib*.so.

BUG=chromium:976360
TEST=USE="asan tpm2 -tpm" FEATURES="test noclean" emerge-amd64-generic cryptohome

Change-Id: Ic08fa25958044f84d6702f3488bf90b53114f556
Reviewed-on: https://chromium-review.googlesource.com/1782232
Tested-by: John L Chen <zuan@chromium.org>
Commit-Ready: John L Chen <zuan@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>