u2fd: Ignore unsupported vendor mode command
Existing builds of cr50 support the vendor mode
commands, but future builds will not. If we get
a NoSuchCommand error, it means we are running
against one of these newer builds, which do not
have a notion of 'mode', and therefore we do not
need to set it. Ignore such errors if we encounter
them.

BUG=b:133275654
TEST=test register in g2f and u2f mode locally on soraka,
     with old and new cr50 firmware

Change-Id: Ice2428fd994f8164865c0ebc292d07a3c4948718
Reviewed-on: https://chromium-review.googlesource.com/1792667
Tested-by: Louis Collard <louiscollard@chromium.org>
Commit-Ready: Louis Collard <louiscollard@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Wei-Cheng Xiao <garryxiao@chromium.org>