authpolicy: Handle POLLERR gracefully
When performing pipe IO for launched child processes, assume that
POLLERR on ouptut pipes means that the pipe should not be written to
anymore because the child process is not interested in its stdin (e.g.
because it has already exited).

POLLERR can be legitimately generated e.g. for pipes that could be
written to but have no more readers. This can happen to the child
process' stdin if the child process doesn't wait for any input and
exits before the call to poll().

To test this handling, added a unit test
ProcessExecutorTest.ChildProcessAlreadyExited which only starts
performing pipe IO when the child process has already exited.

BUG=chromium:997728
TEST=cros_run_unit_tests --board=${BOARD} --packages authpolicy

Change-Id: Icea5925653ca6cc14b4c76006f26e09bf0ad7c47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1806594
Tested-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>