vm_tools: concierge: do not resize disks with fixed_size xattr
Check whether the user.crostini.fixed_size xattr exists on the disk
image before attempting to resize it.  This xattr is not set anywhere
yet, but it will be used to indicate that a VM's disk size has been
chosen by the user and should not be automatically extended.

Manual testing steps (on host):
- `setfattr -n user.crostini.fixed_size -v 1 <termina image path>`
- boot VM and verify disk resize is not attempted by reading logs

BUG=chromium:858815
TEST=Manual test (see above)

Change-Id: I082f76a59e0a0cad9c065bb160cb24d34868311c
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1904827
Reviewed-by: Stephen Barber <smbarber@chromium.org>