arc: networkd: Add patchpanel server and client
This patch introduces DBus server and client implementations for the new
patchpanel service. The server is implemented in Manager which needs to
first be aware of ARCVM starting/stopping (similar to ARC++) and can
forward the applicable calls to the ARCVM service implementation which
will be introduced in a subsequent patch.

BUG=b:133707144
BUG=b:133193765
TEST=Verified as part of this dependent patch chain that using the
client correctly sends and receives messages over DBus handled by the
server.

Change-Id: I1fe4a0ac0bcabe1ab6926a46dbb5d6beb2e18b6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1885532
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>