vm_tools: concierge: share kCrosvmBin definition
kCrosvmBin is already defined in vm_util.h; remove the duplicate copy in
the RunCrosvmCommand function.

Additionally, move the definition to vm_util.cc instead of vm_util.h:
putting the definition of kCrosvmBin in the header means that it will be
duplicated in each file that includes it.  Move the definition to the
.cc file instead so that the single definition can be shared by all
users.

BUG=None
TEST=`strings /build/$BOARD/usr/bin/vm_concierge | grep /usr/bin/crosvm`

Change-Id: I6821d7036d12dec9e483a29830a878889a2ee0e3
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1808350
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>