vm_tools: concierge: Start ARCVM with less CPUs
Previously, the VM was started with '--cpus N' where N is the number
of logical CPU cores on the host (both online and offline). Since
Chrome OS recently disabled hyper threading, on most Intel boards,
usually half of the cores are offline. This CL changes N to the
number of logical cores _online_ on the host.

This is not a final solution. We need to move the logic to Chromium
before we start launching Concierge on login screen. For more
details, see b/139752657#comment3.

BUG=b:139752657
TEST=ARCVM starts with '--cpus 2' on eve.

Change-Id: I412187897c800b006074964eed06318cc2f63c2b
Reviewed-on: https://chromium-review.googlesource.com/1809654
Tested-by: Yusuke Sato <yusukes@chromium.org>
Commit-Ready: Yusuke Sato <yusukes@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>