cryptohome: Improve ResetLECredentials
ResetLECredentials tries to go through expensive authenticate operation
before reseting the low entropy keyset. This is done to reset the number
of wrong attempts of authentication with LE credentials. But many times
that number is zero, so no reset is necessary. In those cases we can avoid
the authentication operation and skip the reset step. This CL changes the
code to achieve that.

BUG=chromium:912126
TEST=Unit tests

Change-Id: I3efbffb83235ba14a760d0cba48cf89053fce11f
Reviewed-on: https://chromium-review.googlesource.com/1390133
Tested-by: Victor-Gabriel Savu <vsavu@google.com>
Commit-Ready: Victor-Gabriel Savu <vsavu@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>