common-mk: Resolve linter warning by removing unused code.
Linter detects use of basestring as a warning because it is not
supported in Python 3.
There is no code that passes a single string or a byte array to the
function as |varnames|. Therefore we can just remove the lines.

BUG=chromium:1005203
TEST=build_packages

Change-Id: Icbe255c278841b018fdac264c1d943b102c91778
Reviewed-on: https://chromium-review.googlesource.com/1810289
Tested-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Commit-Ready: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>