debugd: Add |num_cores_disabled| output param to SetSchedulerConfigurationV2
This allows Chromium to know how many logical CPU cores are disabled.
On eve (an x86_64 device), for example, the method returns 2 when
the pref is set to conservative (the default), and 0 when it is
performance. On non x86_64 devices, it's always 0.

Since this CL only adds a new output parameter the the D-Bus method,
existing Chromium binaries can still talk to the new debugd.

BUG=b:139752657
TEST=FEATURES=test emerge-$BOARD chromeos-base/debugd
TEST=Boot eve, verify the method returns '2'. Kill Chrome browser
 process, wait for Chrome to re-launch, verify the method returns
 '2' again.

Cq-Depend: chromium:1815568
Change-Id: I31a2da7f1c0a8bf51aeece8f87d713c75d460145
Reviewed-on: https://chromium-review.googlesource.com/1814889
Tested-by: Yusuke Sato <yusukes@chromium.org>
Commit-Ready: Yusuke Sato <yusukes@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>