chromeos-config: cleanup redundant x86/ARM init code
This refactors the initialization functions of CrosConfig to no longer
have separated implementations for x86 and ARM, treating both as a
CrosConfigIdentity, and relies on the underlying CrosConfigIdentityX86
or CrosConfigIdentityArm to handle the architecture-dependant
sections.

The (many) Init functions of CrosConfig have been simplified down to
two:

* Init
* InitForTest

An alias of InitModel to Init is provided, to be removed after
everything else in platform2 no longer calls InitModel (solved in
child CLs).

BUG=b:140909519,chromium:999401,chromium:999403
TEST=sudo FEATURES=test emerge chromeos-config-tools

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I1982af049844f295f7a96734056a23e48d9b5b2f
Reviewed-on: https://chromium-review.googlesource.com/1809878
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>