attestation: adaptive key decoding EncryptIdentityCredential
Because of incomplete migration of EK for TPM1.2 case, the factory
method fails to decode EK in the mentioned function. To make sure
what used to work continues to work, we decode the RSA depending on the
TPM version.

BUT=chromium:942487
BUG=b:141344933
TEST=unittest for TPM1.2, TPM2.0
TEST=attestation_client verify_attestation for TPM1.2 case

Change-Id: Ie790b6d04e16ddcb3681a25ba4c3a3811155eb73
Reviewed-on: https://chromium-review.googlesource.com/1815922
Tested-by: Leo Lai <cylai@google.com>
Commit-Ready: Leo Lai <cylai@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>