shill: ethernet: fix check in EthernetProvider::DeregisterService()
If DeregisterService() is called multiple times in quick succession
(before SortServicesTask is run) and not according to the service
sort order, the expression in the if statement in DeregisterService()
will return false when it should return true and the last remaining
service will be deregistered when it shouldn't. This will cause the
CHECK in RefreshGenericEthernetService to fail when SortServicesTask
is eventually run.

The if statement is meant to catch the case in which there's a single
EthernetService left that is being deregistered. However, it won't be
true when the "primary" EthernetService was deregistered first and the
remaining services haven't been sorted yet (thus moving the
ethernet_any pointer from the old "primary" service to the old
"secondary" service).

BUG=chromium:1006026
TEST=unittests
TEST=plugged two Ethernet adapters in and unplugged/plugged in both as
     closely as possible. Did a couple suspend-resumes as well.
TEST=network_WiFi_SuspendStress.11g

Change-Id: If23bdf40b3e6808b3331bb5a508ca1210fe843de
Reviewed-on: https://chromium-review.googlesource.com/1817229
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Ready: Matthew Wang <matthewmwang@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>