power: Notify powersupply observers on battery percentage change.
Currently when we receive power_supply subsystem udev events we update
the observers only if :
   1. external_power status change (one of USB, AC, DISCONNECTED).
   2. battery_state changed (one of FULL, CHARGING, DISCHARGING,
      NOT_PRESENT)
   3. If connected sources are equal.
This code seems to be added with the idea that udev events are sent only
when state changes but not when battery percentage changes. On receving
udev events we also reset one short poll timer that polls the battery
percentage every 30 secs. This means if we receive a udev event due to
battery percentage change at a frequency faster than 30 secs, we end up
not notifying the clients about the change in the battery percentage.
This means ash tray will show same percentage unless there is a power
source change.  This one line change fixes the same bug.

BUG=b:141472256
TEST=unit test & build pass

Change-Id: I1af2e6291e6adbd2ea09b8f0bfaaa6221cdc2547
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1820117
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>