vm_tools: sommelier: Support for active window requests.
We add support for applications to request window activation. One
legitimate us of this request is to foreground a window if the user e.g.
opens a file in a pre-existing window.

There are two common ways for windows to request this, either directly
(via NET_ACTIVE_WINDOW) or indirectly (setting the urgency flag in
their NET_WM_HINTS property). The former will only work if we advertise
that "active window" is supported by this WM.

Currently the request does nothing (except log a message). We will
handle it in a future CL.

BUG=chromium:899587, chromium:978656
TEST=`sudo journalctl | grep "sommelier"`

Change-Id: If9f074d5e5914a8a8e4048e8235d6c4b9b168975
Reviewed-on: https://chromium-review.googlesource.com/1824398
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Ready: Nic Hollingum <hollingum@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>