u2fd: G2fClientTest: Increase timeout
G2fClientTest is flaky on scarlet, likely
hitting a timeout. All code is executing
locally and not blocking, so if this is indeed
happening, seems like it is just because of
CPU load.

The timeout is an arbitrary value, and not part
of the test; increasing it to (hopefully!) fix
flakiness.

BUG=chromium:1003956
TEST=FEATURES=test emerge-soraka g2f_tools

Change-Id: Iaad1b5b6bd925e844ce5635c938c913b6fbfa2d9
Reviewed-on: https://chromium-review.googlesource.com/1824400
Tested-by: Louis Collard <louiscollard@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Wei-Cheng Xiao <garryxiao@chromium.org>