shill: service: Use Disconnecting state
Prior to this change, Service did not use the Disconnecting
state. Instead, a successful Disconnect flow would occur (for the most
part) while the Service was still in whatever state it was when
Disconnect() was first called. At some point the state would be driven
to kStateIdle and the disconnection would be considered complete.

This change causes a Disconnect to trigger a state change to
kStateDiconnecting, which is the state the Service will remain until
the child Device drives the state to kStateIdle. Note that this is a
change which D-Bus clients will see. arc-networkd ignores
non-connected Services, Concierge only looks at the default Service
and ignores the kStateProperty, and Jetstream does convert the state
to CONNSTATE_DISCONNECT but effectively ignores all states other than
CONNSTATE_ONLINE. Chrome has been modified to not trigger a DCHECK
when seeing the disconnect state (CL:1819630) and the TEST section
covers some basic sanity. Future changes that involve remaining in
kStateDisconnecting for more prolonged periods of time may require
further involvement with Chrome developers to ensure that the UI
works as intended.

In addition, this change provides support for a connection attempt
while in the disconnecting state. In this case, the Connect() call
will return immediately (as long as the Service is connectable) and
will be re-triggered once the disconnection has completed.

BUG=chromium:989361
TEST=All unit tests are passing.
TEST=Modified Ethernet::DisconnectFrom to post a delayed task
     (5000ms), then ran ./connect-service "Ethernet" in the middle of
     a ./disconnect-service "Ethernet" call.
       -In logs: "Service 3: state Disconnecting -> Idle" followed by
                 "Connect to service 3: Triggering delayed Connect"
       -With cellular/wifi enabled, a disconnecting Ethernet that has
        not yet dropped its Connection will still have routing rules,
        but the cellular and wifi rules will be higher priority
        because Manager has resorted Services following the transition
        to kStateDisconnecting.
TEST=Manually connecting and disconnecting to ethernet (unmodified),
     cellular, wifi, and VPN networks through the Chrome UI.

Change-Id: Iaec5e9edc8904eaf999c6c975b5300308e45efb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1829659
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>