shill: service: Prevent a Disconnect when a Service is inactive
The Service D-Bus API states that a failure will be returned if a
Disconnect is attempted on a Service that is not active. However, this
is not currently enforced. This change rectifies the inconsistency.

Note that EthernetService, CellularService, WiFiService (through
WiFi::SetPendingService), PPPoEService, and all VPN drivers follow
flows such that the Service will be in an active state by the time
OnConnect returns.

BUG=None
TEST=All unit tests are passing.
TEST=Manual ./connect-service and ./disconnect-service calls to ensure
     that a disconnect call will immediately fail when the Service is
     not active.

Cq-Depend: chrome-internal:1944300
Change-Id: Iab03983539e6650bf0a72e5a047c459ecc4cf3bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1829658
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>