init: fix incorrect result from GetBlockCount
The GetBlockCount helper function falls back to checking
/sys/class/block/[dev]/size to get a count of blocks if dumpe2fs fails.
However, that interface returns a count of sectors (512 bytes) instead of
blocks (often 4096 bytes). This means that if we do get the block count
from this fallback, our count will be 8 times too large.

This causes an error message when we attempt to write beyond the size of
the file system.

Add block size as a parameter to GetBlockCount, so that we can convert
the sector count into a block count.

BUG=None
TEST=None

Change-Id: I6fde163d8553e36ddddda02a0f244c409eb1ccc9
Reviewed-on: https://chromium-review.googlesource.com/1834428
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Commit-Ready: Fletcher Woodruff <fletcherw@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>