hammerd: Introduce the unified build support
To adopt the unified build, we use cros_config to query the
target board name and its configuration for the specific model.
To maintain backwards compatibility with the old style, we
check the ec-image-name, if we can't find it, then just fall
back to the old style.

BUG=b:139038006
TEST=checked the hammerd.log

Change-Id: I7fcc899e966de18be9ebff8f270b8d4dbfb1d9cf
Reviewed-on: https://chromium-review.googlesource.com/1789323
Tested-by: Zhuohao Lee <zhuohao@chromium.org>
Commit-Ready: Zhuohao Lee <zhuohao@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>