init: Check if TPM firmware update is done when deciding to clobber
When TPM firmware update is initiated, TPM owner is cleared but the files
should be preserved. To mark that, preservation_request file is created.

This change checks the existence of preservation_request file and avoids
clobbering if that's the case even if TPM is not owned.

BUG=chromium:1011246
TEST=Perform TPM firmware update on enrolled device and make sure it
doesn't clobber on reboot.

Change-Id: I2762b347bc4b3cffd8fcc1dc595abcd8f5c214dd
Reviewed-on: https://chromium-review.googlesource.com/1846272
Tested-by: Igor <igorcov@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>