shill: pppoe: Get exit codes from pppd
pppd has an option called max_fail. When max_fail > 0, pppd will try
to connect to the peer max_fail times before exiting with an
informative exit code. While PPPoEService does set this value, it also
considers any disconnect from pppd to be a failure and shuts down pppd
when this occurs.

In this way, the code essentially uses max_fail=1, but in a roundabout
way that doesn't involve pppd being aware of the fact. This CL changes
the behavior so that both shill and pppd have coherent views of how
connections should be made, allowing for pppoe code to use the error
code returned from pppd in order to determine what kind of failure, if
any, occurred.

This change is very similar to what was done for cellular in CL:1464146.

BUG=chromium:929781
TEST=`USE="pppoe" cros_workon_make shill --board ${BOARD} --test`

Change-Id: I1052799551e011686c6aefe6f71cc68cbaf62923
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1857179
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>