arc: network: Remove "stateful" multicast forwarder listener
According to mDNS and SSDP documentations, "stateful" multicast
should be replied with unicast. Because of that and the fact that
packets are being sent to ARC (go/arc++inbound), there is no need
to listen on the physical interface.

This patch introduces a behavioral change of:
- Ingress legacy mDNS and SSDP ("stateful") will be forwarded to
  guest.
(Used to be that only egress -> followed by listening to unicasts).

BUG=b:140907324
TEST=Manual (mDNS is working, SSDP and legacy mDNS is forwarded)
TEST=Join and host Expeditions

Change-Id: Idc9bb2c6cd2f375fa8b9157a9f06d6d2df8562c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1848136
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Auto-Submit: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>